AshDoubleEntry.ULID (ash_double_entry v1.0.6)

An Ash type for ULID strings.

Types

t()

@type t() :: <<_::208>>

A hex-encoded ULID string.

Functions

bingenerate(timestamp \\ System.system_time(:millisecond))

Generates a binary ULID.

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

bingenerate_last(timestamp \\ System.system_time(:millisecond))

Generates a binary ULID.

Do not use this for storage, only for generating comparators, i.e "balance as of a given ulid".

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

cast_input(value, constraints)

Casts a string to ULID.

cast_stored(bytes, _)

Converts a binary ULID into a Crockford Base32 encoded string.

dump_to_native(encoded, _)

Converts a Crockford Base32 encoded ULID into a binary.

generate(timestamp \\ System.system_time(:millisecond))

Generates a Crockford Base32 encoded ULID.

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

generate_last(timestamp \\ System.system_time(:millisecond))

Generates a Crockford Base32 encoded ULID, guaranteed to sort equal to or after any other ULID generated for the same timestamp.

Do not use this for storage, only for generating comparators, i.e "balance as of a given ulid".

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

storage_type()

The underlying schema type.